chore: rename component validator to validator decorator
