+++ src/core/validatorDecorator.js
