+++ tests/integration/inject.js
@@ -1,7 +1,7 @@
 import { mount, createLocalVue } from '@vue/test-utils';
 import flushPromises from 'flush-promises';
 import VeeValidate from '@/index';
-import ScopedValidator from '@/core/componentValidator';
+import ScopedValidator from '@/core/validatorDecorator';
 import TestComponent from './components/Inject';
 import BuiltInsTestComponent from './components/BuiltIn';
 import ChildInject from './components/stubs/ChildWithParentValidatorInjection';
